lab3/requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
opencv-python
pillow --only-binary :all:
matplotlib
pyyaml
pandas
tqdm
scipy


# python -m venv env 
# source env/bin/activate 
# python -m pip install --upgrade pip 
# pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu118 --verbose
# cuda 11.8 needs compatible with GPU